+++ SKVS.Server/Repository/IWarehouseOrderRepository.cs
@@ -5,6 +5,7 @@ namespace SKVS.Server.Repository
     public interface IWarehouseOrderRepository
     {
         Task<IEnumerable<WarehouseOrder>> GetAllAsync();
+        Task<IEnumerable<WarehouseOrder>> GetAllSelectedAsync(List<int> orderIds);
         Task<WarehouseOrder?> GetByIdAsync(int id);
         Task AddAsync(WarehouseOrder order);
         Task UpdateAsync(WarehouseOrder order);
