+++ skvs.client/src/main.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router';
-import App from './App.jsx';
+import Router from './Router.jsx';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
 	<React.StrictMode>
 		<BrowserRouter>
-			<App />
+			<Router />
 		</BrowserRouter>
 	</React.StrictMode>
 );
